Merge PJT6 into master branch

See merge request nts-2020-2nd-team5/jeon_kyung_yoon!31

+++ reservation/src/main/java/kr/or/connect/reservation/dao/ProductImageDao.java
@@ -24,7 +24,7 @@ public class ProductImageDao {
 		this.jdbc = new NamedParameterJdbcTemplate(dataSource);
 	}
 
-	public List<ProductImage> selectAllUsingProductId(Integer productId) {
+	public List<ProductImage> selectAllUsingProductId(Long productId) {
 		Map<String, ?> params = Collections.singletonMap("productId", productId);
 		return jdbc.query(SELECT_ALL_USING_PRODUCT_ID, params, rowMapper);
 	}
